add create cart

+++ src/services/cart.service..ts
@@ -1,6 +1,7 @@
 import { Service } from 'typedi';
 import Cart from '../db/models/Cart';
 import { CartRepository } from '../db/repositories';
+import { CreateCartDto } from './../controllers/cart/dtos/create-cart.dto';
 
 @Service()
 class CartService {
@@ -8,6 +9,23 @@ class CartService {
 
   public async getAll(user: string): Promise<Cart[]> {
     return await this.cartRepository.getAllCarts({ user });
+  }
+
+  public async createCart(
+    { productId, qty }: CreateCartDto,
+    user: string
+  ): Promise<Cart> {
+    const existingCart = await this.findByProductId(productId, user);
+
+    if (existingCart) {
+      return await this.updateCartQuantity(productId, qty, user);
+    }
+
+    return await this.cartRepository.createCart({
+      productId: productId,
+      qty: qty,
+      user
+    });
   }
 
   public async findByProductId(
@@ -21,6 +39,10 @@ class CartService {
     productId: string,
     user: string
   ): Promise<void> {
+    const currentCart = await this.findByProductId(productId, user);
+    if (!currentCart) {
+      throw Error('No cart found for the product');
+    }
     await this.cartRepository.removeFromCartByProductId(productId, { user });
   }
